check version when restore cofnig

+++ cfgmgr/files/cfgmgr_71xx/sysbackup.sh
@@ -31,11 +31,20 @@ restore() {
 	timeout -t 10 openssltar.sh untar $inpath $recover_dir wjrc0409
 	test $? -eq 0 || err_exit "openssltar.sh tar fail"
 
-	sysupgrade --restore-backup $recover_dir/cfg.tar.tgz
-	test $? -eq 0 || err_exit "sysupgrade --restore-backup fail"
+	s1=`cat $recover_dir/openwrt_release | grep DISTRIB_DESCRIPTION`
+	test $? -eq 0 || err_exit "read $recover_dir/openwrt_release fail"
+	s2=`cat /etc/openwrt_release | grep DISTRIB_DESCRIPTION`
+	test $? -eq 0 || err_exit "read /etc/openwrt_release fail"
+	if [ "$s1" == "$s2" ]; then 
+		sysupgrade --restore-backup $recover_dir/cfg.tar.tgz
+		test $? -eq 0 || err_exit "sysupgrade --restore-backup fail"
 
-	rm -rf $recover_dir
-	reboot
+		rm -rf $recover_dir
+		echo "result:ok"
+		exit 0
+	fi
+	echo "result:$s1 <======> $s2"
+	exit 1
 }
 
 case $1 in
